[API,BUILD,TEST] Narrow public header dependencies
Forward-declare reference-only types, share ToolInfo independently of CTD,
and keep Parquet implementation includes in the implementation. Compile
eight public headers in separate translation units without unity/PCH.

Part of OpenMS/OpenMS#10083, step 3.
