+++ src/openms/include/OpenMS/METADATA/SpectrumMetaDataLookup.h
@@ -10,16 +10,15 @@
 
 #include <OpenMS/METADATA/SpectrumLookup.h>
 
-#include <OpenMS/FORMAT/FileHandler.h>
 #include <OpenMS/KERNEL/MSSpectrum.h>
 #include <OpenMS/METADATA/ProteinIdentification.h>
-#include <OpenMS/KERNEL/MSExperiment.h>
 #include <OpenMS/METADATA/PeptideIdentificationList.h>
 
 #include <limits> // for "quiet_NaN"
 
 namespace OpenMS
 {
+  class MSExperiment;
 
   /**
     @brief Helper class for looking up spectrum meta data
@@ -338,4 +337,3 @@ namespace OpenMS
   };
 
 } //namespace OpenMS
-
