+++ src/openms/source/FORMAT/QPXCollectionExport.cpp
@@ -15,6 +15,7 @@
 #include <OpenMS/FORMAT/QPXFile.h>
 #include <OpenMS/KERNEL/ConsensusMap.h>
 #include <OpenMS/METADATA/ExperimentalDesign.h>
+#include <OpenMS/METADATA/ProteinIdentification.h>
 #include <OpenMS/METADATA/PeptideIdentificationList.h>
 #include <OpenMS/SYSTEM/File.h>
