+++ src/topp/PhosphoScoring.cpp
@@ -12,6 +12,7 @@
 #include <OpenMS/ANALYSIS/ID/AScore.h>
 #include <OpenMS/DATASTRUCTURES/ListUtils.h>
 #include <OpenMS/METADATA/SpectrumMetaDataLookup.h>
+#include <OpenMS/FORMAT/FileHandler.h>
 
 using namespace OpenMS;
 using namespace std;
